+++ examples/ui-demo/src/hooks/useModularAccountV2Client.ts
@@ -14,20 +14,20 @@ import { Chain, Hex, Address, PrivateKeyAccount } from "viem";
 import { LocalAccountSigner } from "@aa-sdk/core";
 import { privateKeyToAccount } from "viem/accounts";
 
-interface PolicyToken {
-  address: Address;
-  maxTokenAmount: bigint;
-  approvalMode?: "PERMIT" | "NONE";
-  erc20Name?: string;
-  version?: string;
-}
-
 type Client = ModularAccountV2Client<
   AlchemySigner | LocalAccountSigner<PrivateKeyAccount>
 > &
   InstallValidationActions<
     ModularAccountV2<AlchemySigner | LocalAccountSigner<PrivateKeyAccount>>
   >;
+
+interface PolicyToken {
+  address: Address;
+  maxTokenAmount: number;
+  approvalMode?: "PERMIT" | "NONE";
+  erc20Name?: string;
+  version?: string;
+}
 
 // Hook that creates an MAv2 client that can be used for things that
 // @account-kit/react doesn't yet support, such as session keys.